+++ addons/base_setup/views/res_config_settings_views.xml
@@ -210,17 +210,15 @@
 
                         <h2>Integrations</h2>
                         <div class="row mt16 o_settings_container" name="integration">
-                            <div class="col-12 col-lg-6 o_setting_box" id="external_pads_setting">
+                            <div class="col-12 col-lg-6 o_setting_box" id="mail_client_extension_setting">
                                 <div class="o_setting_left_pane">
-                                    <field name="module_pad"/>
+                                    <field name="module_mail_client_extension" />
                                 </div>
                                 <div class="o_setting_right_pane">
-                                    <label for="module_pad"/>
+                                    <label string="Mail Client Extension" for="module_mail_client_extension"/>
+                                    <a href="https://www.odoo.com/documentation/user/crm/optimize/outlook_extension.html" title="Documentation" class="o_doc_link" target="_blank"></a>
                                     <div class="text-muted">
-                                        Use external pads in Odoo Notes
-                                    </div>
-                                    <div class="content-group" attrs="{'invisible': [('module_pad','=',False)]}" id="msg_module_pad">
-                                        <div class="text-warning mt16"><strong>Save</strong> this page and come back here to set up the feature.</div>
+                                        Integrate with mail client plugins
                                     </div>
                                 </div>
                             </div>
@@ -235,6 +233,20 @@
                                         Synchronize your calendar with Outlook
                                     </div>
                                     <div class="content-group" attrs="{'invisible': [('module_microsoft_calendar', '=', False)]}" id="msg_module_microsoft_calendar">
+                                        <div class="text-warning mt16"><strong>Save</strong> this page and come back here to set up the feature.</div>
+                                    </div>
+                                </div>
+                            </div>
+                            <div class="col-12 col-lg-6 o_setting_box" id="external_pads_setting">
+                                <div class="o_setting_left_pane">
+                                    <field name="module_pad"/>
+                                </div>
+                                <div class="o_setting_right_pane">
+                                    <label for="module_pad"/>
+                                    <div class="text-muted">
+                                        Use external pads in Odoo Notes
+                                    </div>
+                                    <div class="content-group" attrs="{'invisible': [('module_pad','=',False)]}" id="msg_module_pad">
                                         <div class="text-warning mt16"><strong>Save</strong> this page and come back here to set up the feature.</div>
                                     </div>
                                 </div>
