+++ addons/crm/models/res_config_settings.py
@@ -39,7 +39,6 @@ class ResConfigSettings(models.TransientModel):
     module_crm_iap_lead = fields.Boolean("Generate new leads based on their country, industries, size, etc.")
     module_crm_iap_lead_website = fields.Boolean("Create Leads/Opportunities from your website's traffic")
     module_crm_iap_lead_enrich = fields.Boolean("Enrich your leads automatically with company data based on their email address.")
-    module_mail_client_extension = fields.Boolean("See and manage users, companies, and leads from our mail client extensions.")
     lead_enrich_auto = fields.Selection([
         ('manual', 'Enrich leads on demand only'),
         ('auto', 'Enrich all leads automatically'),
