+++ addons/crm/views/res_config_settings_views.xml
@@ -121,19 +121,6 @@
                     <div class="row mt16 o_settings_container" name="convert_visitor_setting_container">
                         <div class="col-lg-6 o_setting_box">
                             <div class="o_setting_left_pane">
-                                <field name="module_mail_client_extension"/>
-                            </div>
-
-                            <div class="o_setting_right_pane" id="mail_client_extension">
-                                <label string="Outlook CRM Extension" for="module_mail_client_extension"/>
-                                <a href="https://www.odoo.com/documentation/user/crm/optimize/outlook_extension.html" title="Documentation" class="o_doc_link" target="_blank"></a>
-                                <div class="text-muted">
-                                    Turn emails received in your Outlook mailbox into leads and log their content as internal notes.
-                                </div>
-                            </div>
-                        </div>
-                        <div class="col-lg-6 o_setting_box">
-                            <div class="o_setting_left_pane">
                                 <field name="module_crm_iap_lead_enrich"/>
                             </div>
