+++ addons/crm_iap_lead_enrich/models/__init__.py
@@ -2,5 +2,4 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import crm_lead
-from . import iap_enrich_api
 from . import res_config_settings
