+++ addons/crm_iap_lead_enrich/tests/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import common
 from . import test_lead_enrich
