+++ addons/crm_mail_client_extension/controllers/__init__.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import iap_account
-from . import res_config_settings
-from . import iap_enrich_api
+from . import mail_plugin
