+++ addons/mail_client_extension/__manifest__.py
@@ -6,17 +6,15 @@
     'version': '1.0',
     'category': 'Sales/CRM',
     'sequence': 5,
-    'summary': 'Turn emails received in your Outlook mailbox into leads and log their content as internal notes.',
-    'description': "Turn emails received in your Outlook mailbox into leads and log their content as internal notes.",
-    'website': 'https://www.odoo.com/page/crm',
+    'summary': 'Allows integration with mail plugins.',
+    'description': "Integrate Odoo with your mailbox, get information about contacts directly inside your mailbox, log content of emails as internal notes",
     'depends': [
         'web',
-        'crm',
-        'crm_iap_lead_enrich'
+        'contacts',
+        'iap'
     ],
     'data': [
         'views/mail_client_extension_login.xml',
-        'views/mail_client_extension_lead.xml',
     ],
     'installable': True,
     'application': False,
