+++ addons/mail_client_extension/models/__init__.py
@@ -1,3 +1,5 @@
-from . import crm_lead
-from . import res_partner
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 from . import ir_http
+from . import res_partner
